Inspection rule: validate that `lang` modifier is used only for localized attributes
