+++ resources/i18n/HybrisBundle.properties
@@ -259,6 +259,7 @@ hybris.inspections.impex.ImpexUnknownConfigPropertyInspection.key=[y] Unknown co
 hybris.inspections.impex.ImpexUnknownMacrosInspection.key=[y] Unknown macro ''{0}''
 hybris.inspections.impex.ImpexUnknownAttributeModifierInspection.key=[y] Unknown Attribute modifier ''{0}''
 hybris.inspections.impex.ImpexUnknownTypeModifierInspection.key=[y] Unknown Type modifier ''{0}''
+hybris.inspections.impex.ImpexLanguageModifierIsNotAllowedInspection.key=[y] Lang modifier is not allowed for non-localized attribute ''{0}''
 
 hybris.inspections.ts.AttributeHandlerMustBeSetForDynamicAttribute.key=[y] Attribute handler must be defined for Dynamic Attribute
 hybris.inspections.ts.DeploymentTypeCodeMustBeUnique.key=[y] Deployment type code must be unique
@@ -337,6 +338,9 @@ hybris.inspections.fix.xml.DeleteAttribute=Delete attribute {0}
 hybris.inspections.fix.xml.DeleteSubTag=Delete sub-tag {0}
 hybris.inspections.fix.xml.DeleteTag=Delete tag
 hybris.inspections.fix.psi.NavigateToAnchor=Navigate to existing declaration
+
+hybris.inspections.fix.impex.DeleteModifier=Delete modifier
+hybris.inspections.fix.impex.DeleteModifier.text=Delete modifier ''{0}''
 
 hybris.inspections.fxs.unresolved.attribute.key=[y] Unresolved attribute ''{0}''
 hybris.inspections.fxs.unresolved.type.key=[y] Unresolved type ''{0}''
