+++ src/com/intellij/idea/plugin/hybris/codeInspection/fix/xml/XmlDeleteTagQuickFix.kt
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-package com.intellij.idea.plugin.hybris.codeInspection.fix
+package com.intellij.idea.plugin.hybris.codeInspection.fix.xml
 
 import com.intellij.codeInspection.LocalQuickFix
 import com.intellij.codeInspection.ProblemDescriptor
