+++ src/com/intellij/idea/plugin/hybris/codeInspection/fix/xml/XmlUpdateAttributeQuickFix.java
@@ -16,7 +16,7 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.intellij.idea.plugin.hybris.codeInspection.fix;
+package com.intellij.idea.plugin.hybris.codeInspection.fix.xml;
 
 import com.intellij.codeInspection.LocalQuickFix;
 import com.intellij.codeInspection.ProblemDescriptor;
@@ -25,21 +25,25 @@ import com.intellij.idea.plugin.hybris.common.utils.HybrisI18NBundleUtils;
 import com.intellij.openapi.project.Project;
 import com.intellij.psi.PsiElement;
 import com.intellij.psi.xml.XmlAttribute;
+import com.intellij.psi.xml.XmlElement;
 import com.intellij.psi.xml.XmlTag;
 import com.intellij.util.PsiNavigateUtil;
 import org.jetbrains.annotations.NotNull;
 
-public class XmlAddAttributeQuickFix implements LocalQuickFix {
+public class XmlUpdateAttributeQuickFix implements LocalQuickFix {
 
     private final String myFixName;
     private final String myAttributeName;
+    private final String myAttributeValue;
 
-    public XmlAddAttributeQuickFix(
-        final String attributeName
+    public XmlUpdateAttributeQuickFix(
+        final String attributeName,
+        final String attributeValue
     ) {
 
-        myFixName = HybrisI18NBundleUtils.message("hybris.inspections.fix.xml.AddAttribute", attributeName);
+        myFixName = HybrisI18NBundleUtils.message("hybris.inspections.fix.xml.UpdateAttribute", attributeName, attributeValue);
         myAttributeName = attributeName;
+        myAttributeValue = attributeValue;
     }
 
     @NotNull
@@ -54,14 +58,22 @@ public class XmlAddAttributeQuickFix implements LocalQuickFix {
 
         if (currentElement instanceof XmlTag) {
             final XmlTag currentTag = (XmlTag) currentElement;
-            final XmlAttribute xmlAttribute = currentTag.setAttribute(myAttributeName, "");
-            navigateIfNotPreviewMode(descriptor, xmlAttribute.getValueElement());
+            final XmlAttribute xmlAttribute = currentTag.setAttribute(myAttributeName, myAttributeValue);
+            navigateIfNotPreviewMode(descriptor, xmlAttribute);
+        } else if (currentElement instanceof XmlAttribute) {
+            final XmlAttribute xmlAttribute = (XmlAttribute) currentElement;
+            xmlAttribute.setValue(myAttributeValue);
+            navigateIfNotPreviewMode(descriptor, xmlAttribute);
+        } else if (currentElement instanceof XmlElement && currentElement.getParent() instanceof XmlAttribute) {
+            final XmlAttribute xmlAttribute = (XmlAttribute) currentElement.getParent();
+            xmlAttribute.setValue(myAttributeValue);
+            navigateIfNotPreviewMode(descriptor, xmlAttribute);
         }
     }
 
-    private void navigateIfNotPreviewMode(final ProblemDescriptor descriptor, final PsiElement psiElement) {
+    private void navigateIfNotPreviewMode(final ProblemDescriptor descriptor, final XmlAttribute xmlAttribute) {
         if (descriptor instanceof ProblemDescriptorBase) {
-            PsiNavigateUtil.navigate(psiElement);
+            PsiNavigateUtil.navigate(xmlAttribute);
         }
     }
 }
