+++ src/com/intellij/idea/plugin/hybris/codeInspection/rule/typeSystem/TSUseOfUnoptimizedAttributesIsNotRecommended.kt
@@ -18,7 +18,7 @@
 
 package com.intellij.idea.plugin.hybris.codeInspection.rule.typeSystem
 
-import com.intellij.idea.plugin.hybris.codeInspection.fix.XmlDeleteAttributeQuickFix
+import com.intellij.idea.plugin.hybris.codeInspection.fix.xml.XmlDeleteAttributeQuickFix
 import com.intellij.idea.plugin.hybris.system.type.model.Items
 import com.intellij.idea.plugin.hybris.system.type.model.Modifiers
 import com.intellij.idea.plugin.hybris.system.type.model.modifiers
